refactor: organize deployment access tab files
