+++ web/app/(commonLayout)/deployments/[appInstanceId]/access/page.tsx
@@ -1,4 +1,4 @@
-import { AccessTab } from '@/features/deployments/detail/access-tab/access-tab'
+import { AccessTab } from '@/features/deployments/detail/access-tab'
 
 export default function InstanceDetailAccessPage() {
   return <AccessTab />
