+++ web/app/(commonLayout)/deployments/[appInstanceId]/api-tokens/page.tsx
@@ -1,4 +1,4 @@
-import { DeveloperApiTab } from '@/features/deployments/detail/access-tab/developer-api-tab'
+import { DeveloperApiTab } from '@/features/deployments/detail/access-tab/developer-api'
 
 export default function InstanceDetailApiTokensPage() {
   return <DeveloperApiTab />
