+++ web/features/deployments/detail/access-tab/components/endpoint.tsx
