+++ web/features/deployments/detail/access-tab/developer-api/created-token-dialog.tsx
@@ -6,7 +6,7 @@ import { Dialog, DialogCloseButton, DialogContent, DialogDescription, DialogTitl
 import { toast } from '@langgenius/dify-ui/toast'
 import { useClipboard } from 'foxact/use-clipboard'
 import { useTranslation } from 'react-i18next'
-import { CopyPill } from './common'
+import { CopyPill } from '../components/endpoint'
 
 function buildCurlExample(apiUrl: string, token: string) {
   return `curl -X POST '${apiUrl}' \\
