+++ web/features/deployments/detail/access-tab/developer-api/section.tsx
@@ -13,15 +13,15 @@ import { useAtomValue } from 'jotai'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { consoleQuery } from '@/service/client'
-import { DeploymentEmptyState, DeploymentStateMessage } from '../../components/empty-state'
-import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
-import { DeveloperApiDocsDrawer } from './api-docs-drawer'
+import { DeploymentEmptyState, DeploymentStateMessage } from '../../../components/empty-state'
+import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
+import { CopyPill } from '../components/endpoint'
+import { developerApiSettingsQueryAtom } from '../state'
 import { ApiKeyGenerateMenu } from './api-key-generate-menu'
 import { ApiKeyList } from './api-key-list'
-import { CopyPill } from './common'
-import { CreatedApiTokenDialog } from './developer-api-created-token-dialog'
-import { DeveloperApiSkeleton } from './developer-api-skeleton'
-import { developerApiSettingsQueryAtom } from './state'
+import { CreatedApiTokenDialog } from './created-token-dialog'
+import { DeveloperApiDocsDrawer } from './docs-drawer'
+import { DeveloperApiSkeleton } from './skeleton'
 
 type CreatedApiToken = {
   appInstanceId: string
