+++ web/features/deployments/detail/access-tab/permissions/__tests__/permissions.spec.tsx
@@ -4,10 +4,10 @@ import { AccessMode, AccessSubjectType } from '@dify/contracts/enterprise/types.
 import { fireEvent, render, screen } from '@testing-library/react'
 import { createStore, Provider as JotaiProvider } from 'jotai'
 import { describe, expect, it, vi } from 'vitest'
-import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
-import { EnvironmentPermissionRow } from '../permissions'
-import { AccessPermissionsSection } from '../permissions-section'
-import { accessSettingsQueryAtom } from '../state'
+import { deploymentRouteAppInstanceIdAtom } from '../../../../route-state'
+import { accessSettingsQueryAtom } from '../../state'
+import { EnvironmentPermissionRow } from '../environment-permission-row'
+import { AccessPermissionsSection } from '../section'
 
 const mockMutate = vi.hoisted(() => vi.fn())
 const mockUseAtomValue = vi.hoisted(() => vi.fn())
