+++ web/features/deployments/detail/access-tab/permissions/section.tsx
@@ -4,11 +4,11 @@ import type { EnvironmentAccessPolicy } from '@dify/contracts/enterprise/types.g
 import { useAtomValue } from 'jotai'
 import { useTranslation } from 'react-i18next'
 import { SkeletonRectangle } from '@/app/components/base/skeleton'
-import { DeploymentEmptyState, DeploymentStateMessage } from '../../components/empty-state'
-import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
-import { Section } from '../components/section'
-import { EnvironmentPermissionRow } from './permissions'
-import { accessSettingsQueryAtom } from './state'
+import { DeploymentEmptyState, DeploymentStateMessage } from '../../../components/empty-state'
+import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
+import { Section } from '../../components/section'
+import { accessSettingsQueryAtom } from '../state'
+import { EnvironmentPermissionRow } from './environment-permission-row'
 
 const ACCESS_PERMISSIONS_SKELETON_KEYS = ['production', 'staging', 'development']
